+++ src/components/MainHeader/index.ts
@@ -1,0 +1,1 @@
+export { default as MainHeader } from './MainHeader';
