+++ src/components/index.tsx
@@ -10,6 +10,7 @@ export { Slider } from './Slider';
 export { GameListCard } from './GameListCard';
 export { Navigation } from './Navigation';
 export { HomeClient } from './HomeClient';
+export { MainHeader } from './MainHeader';
 export { default as Label } from './Label';
 export { default as Input } from './Input';
 export { default as Logo } from './Logo';
